requirements.txt
google-adk>=1.23.0
google-cloud-bigquery>=3.11.0
google-auth>=2.22.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.38.0
requests>=2.31.0
pydantic>=2.4.0
